Refactor scale class to be more generic and allow more scale implemetnation
