+++ src/scale.h
@@ -1,12 +1,28 @@
 #ifndef SCALE_H_
 #define SCALE_H_
 
+#include <FreeRTOS.h>
 #include <stdint.h>
 
 #include "app.h"
 #include "http_rest.h"
+#include <semphr.h>
 
-#define EEPROM_SCALE_DATA_REV                     1              // 16 byte 
+#define EEPROM_SCALE_DATA_REV                     2              // 16 byte 
+
+
+typedef struct {
+    uint32_t baudrate;
+} scale_serial_params_t;
+
+
+// Abstracted base class
+typedef struct {
+    // Basic functions
+    void (*read_loop_task)(void *self);
+    void (*write)(char * cmd, size_t len);
+} scale_handle_t;
+
 
 
 typedef enum {
@@ -14,27 +30,46 @@ typedef enum {
     SCALE_UNIT_GRAM = 1,
 } scale_unit_t;
 
+typedef enum {
+    SCALE_DRIVER_AND_FXI = 0,
+} scale_driver_t;
+
+
 typedef struct {
     uint16_t scale_data_rev;
     scale_unit_t scale_unit;
-} __attribute__((packed)) eeprom_scale_data_t;
+    scale_driver_t scale_driver;
+    scale_serial_params_t scale_serial_params;
+} eeprom_scale_data_t;
 
 
+
+typedef struct {
+    eeprom_scale_data_t persistent_config;
+    scale_handle_t * scale_handle;
+    SemaphoreHandle_t scale_measurement_ready;
+    SemaphoreHandle_t scale_serial_write_access_mutex;
+    float current_scale_measurement;
+} scale_config_t;
 
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-// Measurement related calls
+// Scale related calls
 bool scale_init();
-void scale_listener_task(void *p);
 float scale_get_current_measurement();
 float scale_block_wait_for_next_measurement();
-
+const char * get_scale_unit_string(bool);
+void set_scale_unit(scale_unit_t scale_unit);
 bool scale_config_save(void);
 
+// REST
+bool http_rest_scale_weight(struct fs_file *file, int num_params, char *params[], char *values[]);
+bool http_rest_scale_config(struct fs_file *file, int num_params, char *params[], char *values[]);
 
+////////////////////////// A&D Scale Implementations /////////////////////////////
 // Key bindings
 void scale_press_re_zero_key();
 void scale_press_print_key();
@@ -45,16 +80,10 @@ void scale_press_on_off_key();
 void scale_display_off();
 void scale_display_on();
 
+
 // Features
 uint8_t scale_calibrate_with_external_weight();
 AppState_t scale_enable_fast_report(AppState_t prev_state);
-
-const char * get_scale_unit_string(bool);
-void set_scale_unit(scale_unit_t scale_unit);
-
-// REST
-bool http_rest_scale_weight(struct fs_file *file, int num_params, char *params[], char *values[]);
-bool http_rest_scale_config(struct fs_file *file, int num_params, char *params[], char *values[]);
 
 
 #ifdef __cplusplus
